refactor: move DropSiteCommand's logic onto Site.drop()
DropSiteCommand was never a real CLI command (no name/add_arguments,
not registered) — just core drop logic misplaced under pilot/commands.
Moved onto Site.drop() and deleted the file; drop_site_task.py and
tasks/callbacks.py now call Site directly instead of importing from
pilot.commands.
