+++ pilot/tasks/callbacks.py
@@ -51,15 +51,16 @@ def _drop_failed_site(bench_root: Path, site_name: str, site_path: Path) -> bool
     if not (site_path / "site_config.json").is_file():
         return True
     try:
-        from pilot.commands.sites.delete import DropSiteCommand
+        from pilot.config.site import SiteConfig
         from pilot.config.toml_store import BenchTomlStore
         from pilot.core.bench import Bench
+        from pilot.core.site import Site
         from pilot.managers.platform import noninteractive_privileges
 
         config = BenchTomlStore.for_bench(bench_root).read()
         bench = Bench(config, bench_root)
         with noninteractive_privileges():
-            DropSiteCommand(bench, site_name).run()
+            Site(SiteConfig(name=site_name, apps=[]), bench).drop()
         return True
     except Exception as exc:
         logging.debug("Site drop callback failed for %s: %s", site_name, exc)
